feat: 스포츠 레벨 관련 Enum 구현
